requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.7
django-autocomplete-light==3.2.9
django-ckeditor==5.3.0
django-el-pagination==3.1.0
django-endless-pagination==2.0
django-froala-editor==2.6.2
django-haystack==2.4.0
django-ipware==1.1.6
django-js-asset==0.1.1
django-multiselectfield==0.1.7
django-social-share==1.1.2
django-tinymce==2.6.0
google-api-python-client==1.6.4
google-auth==1.1.1
google-auth-httplib2==0.0.2
google-auth-oauthlib==0.1.1
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
oauthlib==2.0.6
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
psycopg2==2.7.3
pyasn1==0.3.7
pyasn1-modules==0.1.5
PyJWT==1.5.2
pysolr==3.3.2
python-social-auth==0.3.6
python3-openid==3.1.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sorl-thumbnail==12.3
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.0